Rename JobRunner to ForwardModelRunner
We no longer call the individual steps in each realization forward model for `Jobs`, and the class name should be changed to reflect that. This commit fixes that.
