+++ pyproject.toml
@@ -86,7 +86,7 @@ dependencies=[
 
 [project.scripts]
 ert = "ert.__main__:main"
-"job_dispatch.py" = "_ert_job_runner.job_dispatch:main"
+"job_dispatch.py" = "_ert_forward_model_runner.job_dispatch:main"
 
 [project.urls]
 Repository = "https://github.com/equinor/ert"
