+++ src/_ert_forward_model_runner/reporting/file.py
@@ -5,9 +5,9 @@ import os
 import socket
 import time
 
-from _ert_job_runner.io import cond_unlink
-from _ert_job_runner.reporting.base import Reporter
-from _ert_job_runner.reporting.message import (
+from _ert_forward_model_runner.io import cond_unlink
+from _ert_forward_model_runner.reporting.base import Reporter
+from _ert_forward_model_runner.reporting.message import (
     _JOB_EXIT_FAILED_STRING,
     _JOB_STATUS_FAILURE,
     _JOB_STATUS_RUNNING,
@@ -19,7 +19,7 @@ from _ert_job_runner.reporting.message import (
     Running,
     Start,
 )
-from _ert_job_runner.util import data as data_util
+from _ert_forward_model_runner.util import data as data_util
 
 TIME_FORMAT = "%H:%M:%S"
 logger = logging.getLogger(__name__)
