+++ tests/unit_tests/ensemble_evaluator/ensemble_evaluator_utils.py
@@ -6,7 +6,7 @@ from cloudevents.http import CloudEvent
 
 from _ert.async_utils import new_event_loop
 from _ert.threading import ErtThread
-from _ert_job_runner.client import Client
+from _ert_forward_model_runner.client import Client
 from ert.ensemble_evaluator import Ensemble, identifiers
 from ert.ensemble_evaluator._builder._realization import ForwardModel, Realization
