+++ tests/unit_tests/forward_model_runner/test_file_reporter.py
@@ -3,9 +3,15 @@ import os.path
 
 import pytest
 
-from _ert_job_runner.job import Job
-from _ert_job_runner.reporting import File
-from _ert_job_runner.reporting.message import Exited, Finish, Init, Running, Start
+from _ert_forward_model_runner.job import Job
+from _ert_forward_model_runner.reporting import File
+from _ert_forward_model_runner.reporting.message import (
+    Exited,
+    Finish,
+    Init,
+    Running,
+    Start,
+)
 from ert.constant_filenames import ERROR_file, LOG_file, STATUS_file, STATUS_json
