+++ tests/unit_tests/forward_model_runner/test_job_dispatch.py
@@ -11,9 +11,9 @@ from unittest.mock import mock_open, patch
 import psutil
 import pytest
 
-from _ert_job_runner.cli import _setup_reporters, main
-from _ert_job_runner.reporting import Event, Interactive
-from _ert_job_runner.reporting.message import Finish, Init
+from _ert_forward_model_runner.cli import _setup_reporters, main
+from _ert_forward_model_runner.reporting import Event, Interactive
+from _ert_forward_model_runner.reporting.message import Finish, Init
 from tests.utils import _mock_ws_thread, wait_until
 
 
@@ -81,7 +81,7 @@ else:
     os.chmod("setsid", 0o755)
 
     job_dispatch_script = importlib.util.find_spec(
-        "_ert_job_runner.job_dispatch"
+        "_ert_forward_model_runner.job_dispatch"
     ).origin
     # (we wait for the process below)
     job_dispatch_process = Popen(
@@ -201,7 +201,7 @@ def test_job_dispatch_run_subset_specified_as_parmeter():
     os.chmod("setsid", 0o755)
 
     job_dispatch_script = importlib.util.find_spec(
-        "_ert_job_runner.job_dispatch"
+        "_ert_forward_model_runner.job_dispatch"
     ).origin
     # (we wait for the process below)
     job_dispatch_process = Popen(
@@ -266,11 +266,11 @@ def test_job_dispatch_kills_itself_after_unsuccessful_job(unused_tcp_port):
     port = unused_tcp_port
     jobs_json = json.dumps({"ens_id": "_id_", "dispatch_url": f"ws://localhost:{port}"})
 
-    with patch("_ert_job_runner.cli.os.killpg") as mock_killpg, patch(
-        "_ert_job_runner.cli.os.getpgid"
+    with patch("_ert_forward_model_runner.cli.os.killpg") as mock_killpg, patch(
+        "_ert_forward_model_runner.cli.os.getpgid"
     ) as mock_getpgid, patch(
-        "_ert_job_runner.cli.open", new=mock_open(read_data=jobs_json)
-    ), patch("_ert_job_runner.cli.JobRunner") as mock_runner:
+        "_ert_forward_model_runner.cli.open", new=mock_open(read_data=jobs_json)
+    ), patch("_ert_forward_model_runner.cli.ForwardModelRunner") as mock_runner:
         mock_runner.return_value.run.return_value = [
             Init([], 0, 0),
             Finish().with_error("overall bad run"),
