+++ tests/utils.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING
 import websockets.server
 
 from _ert.threading import ErtThread
-from _ert_job_runner.client import Client
+from _ert_forward_model_runner.client import Client
 from ert.scheduler.event import FinishedEvent, StartedEvent
 
 if TYPE_CHECKING:
